Fix segfault, found by afl
Since `wdg' gets incremented on each loop iteration it would point one
byte beyond the allocated area if a number is present at the end of the
input. This causes a segfault while evaluating the loop condition where
`wdg' gets deferenced. The solution is to compensate by decrementing the
pointer.